refactor navbar links generation from client to server
